Added ManyToMany notation to persist Records of Secuences Activities

+++ src/main/java/com/cespedesz07/actividad/ActividadRepository.java
@@ -2,5 +2,5 @@ package com.cespedesz07.actividad;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActividadRepository extends CrudRepository<Actividad, Long> {
+public interface ActividadRepository extends CrudRepository<Actividad, Integer> {
 }
